Moved Exceptions to Server Side
Read the title
