+++ src/cs414/a5/groupA/monopoly/server/exception/HotelException.java
@@ -1,4 +1,4 @@
-package cs414.a5.groupA.monopoly.shared.exception;
+package cs414.a5.groupA.monopoly.server.exception;
 
 public class HotelException extends Exception{
     public HotelException(String msg) {
