+++ src/cs414/a5/groupA/monopoly/server/exception/HouseException.java
@@ -1,4 +1,4 @@
-package cs414.a5.groupA.monopoly.shared.exception;
+package cs414.a5.groupA.monopoly.server.exception;
 
 /**
  * Created by bradley on 10/17/16.
